test(wcag-techniques): DRYing tests further
All unit tests at this point followed the exact same execution. Gathering that up into a test "builder" for each test case makes maintenance and reading a lot easier.
