+++ packages/wcag-techniques/t17.tmp.json
@@ -1,0 +1,1 @@
+{"type":"wcag-techniques","metadata":{"passed":0,"warning":0,"failed":1,"inapplicable":0},"assertions":{"QW-WCAG-T17":{"name":"Positioning labels to maximize predictability of relationships","code":"QW-WCAG-T17","mapping":"G162","description":"This technique checks the correct position of labels in forms","metadata":{"target":{"element":"form"},"success-criteria":[{"name":"1.3.1","level":"A","principle":"Perceivable","url":"https://www.w3.org/WAI/WCAG21/Understanding/info-and-relationships"},{"name":"3.3.2","level":"A","principle":"Understandable","url":"https://www.w3.org/WAI/WCAG21/Understanding/labels-or-instructions"}],"related":["H44","H71","H65","G131","G167"],"url":"https://www.w3.org/WAI/WCAG21/Techniques/general/G162.html","passed":2,"warning":0,"failed":2,"inapplicable":0,"outcome":"failed","description":"The form field label is empty."},"results":[{"verdict":"passed","description":"The form field has well positioned label.","resultCode":"P1","elements":[{"htmlCode":"<input id=\"input\" type=\"checkbox\">","pointer":"html > body:nth-child(2) > div:nth-child(1) > form:nth-child(1) > span:nth-child(1) > input:nth-child(1)"}],"attributes":[]},{"verdict":"passed","description":"The form field has well positioned label.","resultCode":"P1","elements":[{"htmlCode":"<input id=\"input2\" type=\"number\">","pointer":"html > body:nth-child(2) > div:nth-child(1) > form:nth-child(1) > input:nth-child(7)"}],"attributes":[]},{"verdict":"failed","description":"The form field label is empty.","resultCode":"F2","elements":[{"htmlCode":"<input id=\"input3\" type=\"radio\">","pointer":"html > body:nth-child(2) > div:nth-child(1) > form:nth-child(1) > input:nth-child(9)"}],"attributes":[]},{"verdict":"failed","description":"The form field label is empty.","resultCode":"F2","elements":[{"htmlCode":"<input id=\"input4\" type=\"email\">","pointer":"html > body:nth-child(2) > div:nth-child(1) > form:nth-child(1) > div:nth-child(15) > input:nth-child(1)"}],"attributes":[]}]}}}
